+++ lib/wice/columns/column_rails_datetime_helper.rb
@@ -1,0 +1,54 @@
+# encoding: utf-8
+module Wice
+  module Columns #:nodoc:
+    class ViewColumnRailsDatetimeHelper < ViewColumn #:nodoc:
+
+      include ActionView::Helpers::DateHelper
+      include Wice::Columns::CommonDateDatetimeMixin
+      include Wice::Columns::CommonStandardDateDatetimeMixin
+
+      def chunk_names
+        %w(year month day hour minute)
+      end
+
+      def do_render(params) #:nodoc:
+        '<div class="date-filter">' +
+          select_datetime(params[:fr], include_blank: true, prefix: @name1) + '<br/>' +
+          select_datetime(params[:to], include_blank: true, prefix: @name2) +
+          '</div>'
+      end
+
+      # name_and_id_from_options in Rails Date helper does not substitute '.' with '_'
+      # like all other simpler form helpers do. Thus, overriding it here.
+      def name_and_id_from_options(options, type)  #:nodoc:
+        options[:name] = (options[:prefix] || DEFAULT_PREFIX) + (options[:discard_type] ? '' : "[#{type}]")
+        options[:id] = options[:name].gsub(/([\[\(])|(\]\[)/, '_').gsub(/[\]\)]/, '').gsub(/\./, '_').gsub(/_+/, '_')
+      end
+
+      def has_auto_reloading_calendar? #:nodoc:
+        false
+      end
+
+    end
+
+    class ConditionsGeneratorColumnRailsDatetimeHelper < ConditionsGeneratorColumn  #:nodoc:
+      def generate_conditions(table_alias, opts)   #:nodoc:
+        conditions = [[]]
+        if opts[:fr]
+          conditions[0] << " #{@column_wrapper.alias_or_table_name(table_alias)}.#{@column_wrapper.name} >= ? "
+          conditions << opts[:fr]
+        end
+
+        if opts[:to]
+          conditions[0] << " #{@column_wrapper.alias_or_table_name(table_alias)}.#{@column_wrapper.name} <= ? "
+          conditions << (opts[:to])
+        end
+
+        return false if conditions.size == 1
+
+        conditions[0] = conditions[0].join(' and ')
+        conditions
+      end
+    end
+  end
+end