+++ lib/wice_grid.rb
@@ -16,6 +16,9 @@ require 'wice/wice_grid_spreadsheet.rb'
 require 'wice/wice_grid_serialized_queries_controller.rb'
 require 'wice/columns/column_processor_index.rb'
 require 'wice/columns.rb'
+require 'wice/columns/common_date_datetime_mixin.rb'
+require 'wice/columns/common_standard_helper_date_datetime_mixin.rb'
+require 'wice/columns/common_js_date_datetime_mixin.rb'
 require 'kaminari.rb'
 
 ActionController::Base.send(:helper_method, :wice_grid_custom_filter_params)